move test to NotFollowTest
